requirements.txt
fitz==0.0.1.dev2
langchain==0.3.23
matplotlib==3.7.5
pandas==2.0.3
pdfplumber==0.11.5
Pillow==11.2.1
pytesseract==0.3.13
spacy==3.7.6
streamlit==1.40.1
textblob==0.19.0
